requirements.txt
selenium==4.10.0
webdriver_manager==3.8.6
requests==2.31.0
urllib3==2.0.3
confluent_kafka==2.1.1
kafka-python==2.0.2
python-dotenv==1.0.0
pytz==2023.3
bs4==0.0.1